fix: operator scheduling and E2E test file structure
- Override nodeSelector in operator values to allow scheduling on any node
  (Talos nodes may not have control-plane label set correctly)
- Move TestE2ESelfHealingHA to proper _test.go file for test discovery
- Remove duplicate test function from phase_selfhealing.go
- Remove unused hcloud import from phase_selfhealing.go
